# backend/internal/http/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Arrflix API",
        "contact": {},
        "version": "0.0.1"
    },
    "basePath": "/api",
    "paths": {
...
        "/v1/indexer": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "indexers"
                ],
                "summary": "Save indexer configuration",
                "parameters": [
                    {
                        "description": "Save indexer",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.IndexerInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.IndexerOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/indexer/action/{name}": {
            "post": {
                "tags": [
                    "indexers"
                ],
                "summary": "Perform an action on an indexer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Action name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Action input",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.IndexerDefinition"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {}
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/indexer/test": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "indexers"
                ],
                "summary": "Test unsaved indexer configuration",
                "parameters": [
                    {
                        "description": "Indexer config",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.IndexerInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.IndexerTestResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/indexer/{id}": {
            "delete": {
                "tags": [
                    "indexers"
                ],
                "summary": "Delete indexer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Indexer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/indexer/{id}/test": {
            "post": {
                "tags": [
                    "indexers"
                ],
                "summary": "Test saved indexer",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Indexer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.IndexerTestResult"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/indexer/{id}/toggle": {
            "put": {
                "tags": [
                    "indexers"
                ],
                "summary": "Toggle indexer enable state",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Indexer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.IndexerOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/indexers/testall": {
            "post": {
                "tags": [
                    "indexers"
                ],
                "summary": "Test all indexers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.IndexerBatchTestResult"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/indexers/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "indexers"
                ],
                "summary": "Get indexer by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Indexer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.IndexerOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Capabilities": {
            "type": "object",
            "required": [
                "bookSearchParams",
                "categories",
                "limitsDefault",
                "limitsMax",
                "movieSearchParams",
                "musicSearchParams",
                "searchParams",
                "supportsRawSearch",
                "tvSearchParams"
            ],
            "properties": {
                "bookSearchParams": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Categories"
                    }
                },
                "limitsDefault": {
                    "type": "integer"
                },
                "limitsMax": {
                    "type": "integer"
                },
                "movieSearchParams": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "musicSearchParams": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "searchParams": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "supportsRawSearch": {
                    "type": "boolean"
                },
                "tvSearchParams": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "model.Categories": {
            "type": "object",
            "required": [
                "id",
                "name",
                "subCategories"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "subCategories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Categories"
                    }
                }
            }
        },
...
        "model.FieldInput": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "value": {}
            }
        },
        "model.FieldOutput": {
            "type": "object",
            "required": [
                "name",
                "privacy"
            ],
            "properties": {
                "advanced": {
                    "type": "boolean"
                },
                "helpLink": {
                    "type": "string"
                },
                "helpText": {
                    "type": "string"
                },
                "hidden": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "order": {
                    "type": "integer"
                },
                "privacy": {
                    "type": "string"
                },
                "selectOptions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SelectOption"
                    }
                },
                "selectOptionsProviderAction": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "value": {}
            }
        },
...
        "model.IndexerBatchTestResult": {
            "type": "object",
            "required": [
                "indexer_id",
                "indexer_name",
                "success"
            ],
            "properties": {
                "error": {
                    "type": "string"
                },
                "indexer_id": {
                    "type": "integer"
                },
                "indexer_name": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "model.IndexerDefinition": {
            "type": "object",
            "required": [
                "added",
                "appProfileId",
                "capabilities",
                "configContract",
                "definitionName",
                "description",
                "downloadClientId",
                "enable",
                "encoding",
                "fields",
                "implementation",
                "implementationName",
                "indexerUrls",
                "infoLink",
                "language",
                "legacyUrls",
                "name",
                "presets",
                "priority",
                "privacy",
                "protocol",
                "redirect",
                "sortName",
                "supportsPagination",
                "supportsRedirect",
                "supportsRss",
                "supportsSearch",
                "tags"
            ],
            "properties": {
                "added": {
                    "type": "string"
                },
                "appProfileId": {
                    "type": "integer"
                },
                "capabilities": {
                    "$ref": "#/definitions/model.IndexerCapabilities"
                },
                "configContract": {
                    "type": "string"
                },
                "definitionName": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "downloadClientId": {
                    "type": "integer"
                },
                "enable": {
                    "type": "boolean"
                },
                "encoding": {
                    "type": "string"
                },
                "fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.IndexerField"
                    }
                },
                "implementation": {
                    "type": "string"
                },
                "implementationName": {
                    "type": "string"
                },
                "indexerUrls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "infoLink": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "legacyUrls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "presets": {
                    "type": "array",
                    "items": {}
                },
                "priority": {
                    "type": "integer"
                },
                "privacy": {
                    "type": "string"
                },
                "protocol": {
                    "type": "string"
                },
                "redirect": {
                    "type": "boolean"
                },
                "sortName": {
                    "type": "string"
                },
                "supportsPagination": {
                    "type": "boolean"
                },
                "supportsRedirect": {
                    "type": "boolean"
                },
                "supportsRss": {
                    "type": "boolean"
                },
                "supportsSearch": {
                    "type": "boolean"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "model.IndexerInput": {
            "type": "object",
            "required": [
                "appProfileId",
                "configContract",
                "enable",
                "fields",
                "implementation",
                "name",
                "priority",
                "protocol",
                "redirect"
            ],
            "properties": {
                "appProfileId": {
                    "type": "integer"
                },
                "configContract": {
                    "type": "string"
                },
                "enable": {
                    "type": "boolean"
                },
                "fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.FieldInput"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "implementation": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "priority": {
                    "type": "integer"
                },
                "protocol": {
                    "$ref": "#/definitions/model.Protocol"
                },
                "redirect": {
                    "type": "boolean"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "model.IndexerOutput": {
            "type": "object",
            "required": [
                "added",
                "appProfileId",
                "configContract",
                "definitionName",
                "description",
                "enable",
                "fields",
                "implementation",
                "implementationName",
                "indexerUrls",
                "infoLink",
                "language",
                "legacyUrls",
                "name",
                "priority",
                "privacy",
                "protocol",
                "redirect",
                "sortName",
                "supportsRedirect",
                "supportsRss",
                "supportsSearch",
                "tags"
            ],
            "properties": {
                "added": {
                    "type": "string"
                },
                "appProfileId": {
                    "type": "integer"
                },
                "capabilities": {
                    "$ref": "#/definitions/model.Capabilities"
                },
                "configContract": {
                    "type": "string"
                },
                "definitionName": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "enable": {
                    "type": "boolean"
                },
                "encoding": {
                    "type": "string"
                },
                "fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.FieldOutput"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "implementation": {
                    "type": "string"
                },
                "implementationName": {
                    "type": "string"
                },
                "indexerUrls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "infoLink": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "legacyUrls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "priority": {
                    "type": "integer"
                },
                "privacy": {
                    "type": "string"
                },
                "protocol": {
                    "$ref": "#/definitions/model.Protocol"
                },
                "redirect": {
                    "type": "boolean"
                },
                "sortName": {
                    "type": "string"
                },
                "supportsRedirect": {
                    "type": "boolean"
                },
                "supportsRss": {
                    "type": "boolean"
                },
                "supportsSearch": {
                    "type": "boolean"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "model.IndexerTestResult": {
            "type": "object",
            "required": [
                "success"
            ],
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "model.Protocol": {
            "type": "string",
            "enum": [
                "unknown",
                "usenet",
                "torrent"
            ],
            "x-enum-varnames": [
                "ProtocolUnknown",
                "ProtocolUsenet",
                "ProtocolTorrent"
            ]
        },
...
    }
}